Create Camera, Texture, and VertexBufferLayout classes and complete VertexArray

+++ include/Renderer/VertexBuffer.h
@@ -3,7 +3,7 @@
 class VertexBuffer
 {
 public:
-    VertexBuffer(const void* data, size_t size);
+    VertexBuffer(const void* data, unsigned int size);
     ~VertexBuffer();
 
     void bind() const;
